#400 apple skia image refactor
